Lint Phase 1: resolve bugprone-command-processor (5), bugprone-random-generator-seed (2), bugprone-implicit-widening-of-multiplication-result (3)
command-processor: all 5 are Simulator.cpp's system() calls invoking
external simulation tools (beep/Seq-Gen/rose/gnuplot) via hardcoded,
developer-local paths. Checked (not assumed) who calls into this file:
only CreateSimulatedData.cpp and BootstrapStats.cpp, both offline,
researcher-invoked simulation tools with no untrusted/network input -
every value interpolated into a command string here is a fixed
constant or a numeric parameter from that same researcher's own
invocation. NOLINT with a file-header comment explaining the check,
not a rewrite.

random-generator-seed: fastdist/fastprot's srand(time(nullptr)) default
(used only when the user doesn't pass an explicit --seed) is for
bootstrap resampling, not a security context - predictable seeding
here has no exploitable consequence, and time-seeding by default is
exactly the intended behavior. NOLINT with a one-line explanation.

implicit-widening-of-multiplication-result: bench_primitives.cpp's and
ProtSeqCompare_test.cpp's `20 * 20` (a compile-time constant, nowhere
near overflowing int) and computeTAMURANEIDistance_DNA_b128_String.cpp's
num_level_1*3 (num_level_1 is a small per-level loop-iteration count,
never realistically large enough to overflow int) are both genuinely
safe - no overflow risk in either, verified by checking the actual
value ranges rather than assuming. Fixed with explicit widening casts
per clang-tidy's own suggested fix, since that's free and leaves no
ambiguity for a future reader.

This closes out Phase 1 (correctness-relevant bugprone-* findings):
re-running the audit now shows zero findings in every bugprone-*
category except bugprone-narrowing-conversions (352, Phase 2's
dedicated triage) and bugprone-easily-swappable-parameters (17, Phase
5's API-design work) - both scoped to later phases from the start.

Verified via full rebuild + ctest + RunExamples.sh, all 15 fixtures
byte-identical to expected_output/.

+++ src/fastphylo/dna/computeTAMURANEIDistance_DNA_b128_String.cpp
@@ -296,7 +296,7 @@ DNA_b128_String::computeTAMURANEIDistance(const DNA_b128_String &s1,
   //     sum_with_previous_level(total_sum_del,sum_del_l1, TWO_BIT_MASK, TWO);
   //   }
   b128 sum_ts_l1,  sum_pyrts_l1, sum_tv_l1,  sum_del_l1;
-  PREFETCH_DATA(num_level_1*3);
+  PREFETCH_DATA(static_cast<size_t>(num_level_1)*3);
   for (  ; num_level_1 != 0 ; num_level_1-- ){    
     dist_level_1(sum_ts_l1,  sum_pyrts_l1, sum_tv_l1,  sum_del_l1);
     sum_with_previous_level(total_sum_ts,sum_ts_l1, TWO_BIT_MASK, TWO);
